requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.1.4
pytest
psycopg2-binary
sqlalchemy
gspread
google-auth